🚚[MOVE] Delegate, Enum 코드 다른 파일로 이동(#34)
