+++ Umpah-iOS/Umpah-iOS/Source/Classes/SelectedRangeDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  SelectedRangeDelegate.swift
+//  Umpah-iOS
+//
+//  Created by SHIN YOON AH on 2021/09/28.
+//
+
+import Foundation
+
+protocol SelectedRangeDelegate: class {
+    func didClickedRangeButton()
+    func didClickedStrokeButton()
+}
