+++ Umpah-iOS/Umpah-iOS/Source/Views/Cells/Main/DateTVC.swift
@@ -7,6 +7,9 @@
 
 import UIKit
 
+import Then
+import SnapKit
+
 class DateTVC: UITableViewCell {
     static let identifier = "DateTVC"
