+++ Umpah-iOS/Umpah-iOS/Source/Views/Cells/Main/FilterCVC.swift
@@ -8,6 +8,7 @@
 import UIKit
 
 import Then
+import SnapKit
 
 class FilterCVC: UICollectionViewCell {
     static let identifier = "FilterCVC"
